cleanup and latency measurement

+++ hck_engine.h
@@ -1,3 +1,3 @@
 void processing_thread();
 int connect_to_hck();
-unsigned short execute_check(int fd, const char* addr, const char* port, bool retry = true);
+double execute_check(int fd, const char* addr, const char* port, bool retry = true);
